moved split conversion in interface
